--- a/.travis.yml
+++ b/.travis.yml
@@ -9,6 +9,7 @@ matrix:
         - brew install moreutils
         - brew install shellcheck
         - brew install jq
+        - brew install coreutils
       script: make test
     - os: linux
       dist: trusty
@@ -24,4 +25,5 @@ matrix:
             - moreutils
             - fakeroot
             - jq
+            - coreutils
       script: debuild -uc -us