--- a/share/github-backup-utils/ghe-backup-config
+++ b/share/github-backup-utils/ghe-backup-config
@@ -309,10 +309,10 @@ version() {
 # 2.18.8, and 2.19.3.  So we need to account for this difference here.
 fix_paths_for_ghe_version() {
   ruby -e '
-    version = ARGV.first
-    ghe_remote_version = version[0] == "v" ? version[1..-1] : version # removing "v" literal from version string
-    version_array = ghe_remote_version.split(".").map(&:to_i)
+    version = ARGV.first.gsub(/^v/, "")  # removing any leading "v" from version string
+    version_array = version.split(".").map(&:to_i)
     fail unless version_array.first >= 2
+
     is_modern = version_array[0] > 2 ||
       (version_array[0] == 2 && version_array[1] > 19) ||
       (version_array[0] == 2 && version_array[1] == 19 && version_array[2] >= 3) ||