--- a/test/test-ghe-backup.sh
+++ b/test/test-ghe-backup.sh
@@ -352,7 +352,7 @@ begin_test "ghe-backup backup gist"
     timeout 2 bash -c "
     source \"$TESTS_DIR\"/../share/github-backup-utils/ghe-backup-config;
     GHE_REMOTE_VERSION=2.16.23;
-    seq 1 10000 | sed -e 's/$/ gist/' | parse_paths | grep -c gist"
+    seq 1 10000 | sed -e 's/$/ gist/' | fix_paths_for_ghe_version | grep -c gist"
 )
 end_test
 
@@ -363,6 +363,6 @@ begin_test "ghe-backup backup wiki"
     timeout 2 bash -c "
     source \"$TESTS_DIR\"/../share/github-backup-utils/ghe-backup-config;
     GHE_REMOTE_VERSION=2.16.23;
-    seq 1 10000 | sed -e 's/$/ wiki/' | parse_paths | grep -c '^\.$'"
+    seq 1 10000 | sed -e 's/$/ wiki/' | fix_paths_for_ghe_version | grep -c '^\.$'"
 )
 end_test