--- a/test/test-ghe-backup.sh
+++ b/test/test-ghe-backup.sh
@@ -368,3 +368,45 @@ begin_test "ghe-backup backup wiki"
     "
 )
 end_test
+
+# check fix_paths_for_ghe_version version thresholds
+begin_test "ghe-backup fix_paths_for_ghe_version newer/older"
+(
+    set -e
+
+    # modern versions keep foo/gist as foo/gist
+    for ver in 2.16.23 2.17.14 2.18.8 2.19.3 2.20.0 3.0.0; do
+        echo == $ver, not gist
+        [ "$(bash -c "
+            source \"$TESTS_DIR\"/../share/github-backup-utils/ghe-backup-config;
+            GHE_REMOTE_VERSION=$ver;
+            echo foo/bar | fix_paths_for_ghe_version
+        ")" == "foo" ]
+
+        echo == $ver, gist
+        [ "$(bash -c "
+            source \"$TESTS_DIR\"/../share/github-backup-utils/ghe-backup-config;
+            GHE_REMOTE_VERSION=$ver;
+            echo foo/gist | fix_paths_for_ghe_version
+        ")" == "foo/gist" ]
+    done
+
+    # old versions change foo/gist to foo
+    for ver in 2.0.0 2.15.123 2.16.22 2.17.13 2.18.7 2.19.2; do
+        echo == $ver, not gist
+        [ "$(bash -c "
+            source \"$TESTS_DIR\"/../share/github-backup-utils/ghe-backup-config;
+            GHE_REMOTE_VERSION=$ver;
+            echo foo/bar | fix_paths_for_ghe_version
+        ")" == "foo" ]
+
+        echo == $ver, gist
+        [ "$(bash -c "
+            source \"$TESTS_DIR\"/../share/github-backup-utils/ghe-backup-config;
+            GHE_REMOTE_VERSION=$ver;
+            echo foo/gist | fix_paths_for_ghe_version
+        ")" == "foo" ]
+    done
+
+)
+end_test