#524 introduced a performance regression in `ghe-backup-repositories`.  This PR speeds up the code introduced in #524 by roughly 3000x.  It also fixes a future bug, in which GitHub Enterprise versions like 2.20.x and 3.x.y would have received the deprecated behavior.  Finally, the PR adds performance and correctness tests to verify the changes.

**Performance:** Prior to the PR, the `parse_paths` step took 80 seconds to process 10k lines, due to `parse_paths` forking and executing `awk`, `grep`, and sometimes `dirline` for each line of the file.  At one large customer, this added over three hours to `ghe-backup` invocations, just doing text processing in the shell.  We believe other customers will be affected as well, proportional to the number of repos, wikis, and gists they are backing up.

We eliminated the redundant calls to `version` (which invokes `awk`), which brought the 10k-line time down to 20 seconds.  Replacing the `grep` call with a shell-intrinsic string match brought the time under 10 seconds.  Still not good enough.

So ultimately we rewrote `parse_paths` as an inline Ruby script, to completely eliminate fork and exec calls.  Each 10k lines now takes around 0.03 seconds, which will remove all but ~3 seconds of the 3-hour overhead at the customer who reported the bug.