@piki 2020-01-30
cc coauthors @ewgenius and @oakeyc 
cc @lildue for review as the historical owner of much of this code